style: change i32 to usize

+++ starky/src/constant.rs
@@ -10,7 +10,7 @@ lazy_static::lazy_static! {
     pub static ref OFFSET_2_64: Fr = Fr::from_str("18446744073709551616").unwrap();
     pub static ref OFFSET_2_128: Fr = Fr::from_str("340282366920938463463374607431768211456").unwrap();
     pub static ref OFFSET_2_192: Fr = Fr::from_str("6277101735386680763835789423207666416102355444464034512896").unwrap();
-    pub static ref CHALLENGE_MAP: HashMap<&'static str, i32> = {
+    pub static ref CHALLENGE_MAP: HashMap<&'static str, usize> = {
         let mut m = HashMap::new();
         m.insert("u", 0);
         m.insert("defVal", 1);
